repo/cs9: switch to development repositories
We need more up-to-date packages than what's available in the production
repositories for CS9.  Since it hasn't been released yet, a lot of
packages are still building support for the distro and getting released
frequently, so testing against "production" can cause issues.

Currently, we need the latest osinfo-db for testing in CI and, although
it was released 3 weeks ago and was made available immediately in the
development repos, it is still not available in production.