gateway: increase pagesize of enumeration
Increase the PageSize of the bucket enumeration to 2^16. It is unclear
what it defaults to, since boto3 does not document it. But testing shows
that it might be somewhere close to 100, given how long the operation
takes. Increasing it to 1000 significantly reduces the time it takes to
something like 3s.

Note that AWS currently refused to serve more than 1000, but this is not
set in stone so it might change in the future.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>