repo: add trailing base-url slashes
Make sure base-urls are suffixed with a slash. They represent
directories, so annotate them as such.

In most cases, this does not matter. However, when we want to modify
paths of those urls, the differnt URL modification helpers need a hint
whether the final part of the URL is a file or not, so they know
whether to strip it or not.

Avoid having to hard-code this behavior, and just use trailing slashes
everywhere. A follow-up commit will verify this in CI, this commit just
prepares for the CI to not fail.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>