Print logs of failed snapshot jobs in the action log
If any job which is part of the scheduled batch fails, the GH action
currently prints its ID and a filer for AWS Cloudwatch to find it.
However this means an extra step of logging into AWS and looking for the
log.

Change the GH action to print the logs for the failed job directly in
the GH action log.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>