mark all static repositories as singletons
All EL8 "latest-RHEL-*" composes under rel-eng seem to be static - they
never change. It's hard to find a definite documentation on this matter,
but I think they are simply the "golden" RC composes.

Thus, let's mark them as singleton with today's date in order to make one
last snapshot and then stop making them so we don't waste resources.

The same is true for EL7. For dotNET-1.1-RHEL-7-20161128.0... 😅