github: Check updated Schutzfile's snapshots
The update_schutzfile.py script just substitutes new dates for the old
ones, which can result in incorrect snapshots if one or more of them
have been removed.

This patch runs the check-snapshot script from the $REPO being modified,
checking to make sure all the new snapshots are available.

If they are not, it exits with a 1. The upstream Schutzfile will need to
be changed to remove the missing snapshots (or the script needs to be
modified to leave them at their final date -- but that's a subject for
another patch).