Add support for Fedora ELN repositories
Fedora ELN is basically a rolling, version-less distribution[1].

Thus, I had to modify gen-all-repos.py to work even if the release isn't
there.

gen-repos.py got a new generator just for ELN. I was thinking whether I can
reuse one of the others, but since ELN is so straightforward, I decided to
just create a very tiny specialized one.

[1] /etc/os-release currently specifies VERSION_ID as 40, but platform is just
eln all the time. RPM's %dist macro is versioned, but this version is
arbitrary because it's bumped every time a mass-rebuild is needed, see:
https://fedoraproject.org/wiki/Changes/ELN_Buildroot_and_Compose#Detailed_Description

Thus, I do believe that ELN can be treated as a version-less
distribution.